Update Deck.java
I changed the print deck method so that it doesn't print the blank cards which represent empty space in the deck array. In addition, I've create a deal card method, which currently only moves each element in the array one step to the left and replaces the last element with an empty card. I also wrote a top card method, to return the value of the card which is on top of the deck. I'm hoping that we can use this to determine which card is next to be drawn/delt.